domain current weather models, current weather mappers, weather type icons

+++ app/src/main/java/ru/yotfr/temps/data/remote/model/MainParametersDTO.kt
@@ -4,13 +4,13 @@ import com.google.gson.annotations.SerializedName
 
 data class MainParametersDTO(
     @SerializedName("temp")
-    val temperature: Int,
+    val temperature: Double,
     @SerializedName("feels_like")
-    val apparentTemperature: Int,
+    val apparentTemperature: Double,
     @SerializedName("temp_min")
-    val minimumTemperature: Int,
+    val minimumTemperature: Double,
     @SerializedName("temp_max")
-    val maximumTemperature: Int,
+    val maximumTemperature: Double,
     val pressure: Int,
     val humidity: Int
 )
