+++ app/src/main/java/ru/yotfr/temps/data/remote/model/WindParametersDTO.kt
@@ -3,8 +3,8 @@ package ru.yotfr.temps.data.remote.model
 import com.google.gson.annotations.SerializedName
 
 data class WindParametersDTO(
-    val speed: Int,
+    val speed: Double,
     @SerializedName("deg")
     val directionDegrees: Int,
-    val gust: Int,
+    val gust: Double,
 )
